% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_anomaly_map.R
\name{autoplot_anomaly_map}
\alias{autoplot_anomaly_map}
\title{Wrapper function to calculate and visualize anomaly of indices for selected years}
\usage{
autoplot_anomaly_map(
  dat_grid,
  dat_p,
  index,
  index_args = list(),
  refyears = NULL,
  anomyears,
  NAmaxClim = 20,
  col = "default",
  title = "",
  plot_title = TRUE,
  plot_args = list(),
  output = NULL,
  plotdir,
  plotname = ""
)
}
\arguments{
\item{dat_grid}{Climindvis-object with type = "grid" generated by \code{\link{make_object}}.}

\item{dat_p}{Climindvis-object with type = "p" generated by \code{\link{make_object}}.}

\item{index}{Name of index to be calculated., e.g. index = "dd" or index = "tnn". For list of indices see \code{\link{indices_list}}}

\item{index_args}{List of arguments for index. See \code{\link{indices_list}}.}

\item{refyears}{Optional. Array of reference years for calculating anomaly. If not defined, all years in dataset are used (default).}

\item{anomyears}{Year or array of years for which anomaly is calculated with respect to refyears. If more than one anomyear is given, the mean of these years is calculated and subtracted from the mean of refyears.}

\item{NAmaxClim}{Integer [0,100]. Maximum percentage of years with missing values for calculation of climatology. If exceeded value for respective grid point/station is set to NA. Default = 20 ( = 20"\%").}

\item{col}{Array of colors for plotting of length 3 or longer (must be a valid arguments to \code{\link[grDevices]{col2rgb}}). Length of array is adjusted to number of levels. If not provided, package default colors for index anomaly are used. Default="default".}

\item{title}{Optional. Character string. If provided added above the plot in addition to default information from the respective function (e.g. index name/aggregation/time period/...). Default = "", no additional title is written.}

\item{plot_title}{Logical. Add plot title to graphic. Default = TRUE.}

\item{plot_args}{Optional. List of arguments passed to plotting function, see \code{\link{plot_args_map}}.}

\item{output}{Format of image to be plotted ("png", "jpeg", "tiff", "pdf","dev.new").  Default = NULL, output to null device (getOption("device") will show you the null device. If you are not using RStudio, NULL and dev.new will give the same output).}

\item{plotdir}{Character string of path of directory for saving plots. Ignored if  \emph{output} is NULL or "dev.new".}

\item{plotname}{Optional. Character string. First part of plot name followed by character string containing index name/aggregation/time period/... depending on the autoplot function. Default = "".}
}
\description{
This function calls the function  \code{\link{calc_index}} and plots the anomaly of the mean of the index values of defined years against the mean of a set of reference years in a map. It can be used for either gridded or station data or a combination of both. \cr
}
\section{Output}{

This function returns one plot per temporal aggregation whch is covered by the data( e.g. if aggt="seasonal" it would return 4 graphics, one for each season).

If output is not NULL or "dev.new" the graphics are directly saved in the user specified directory (\emph{plotdir}) with the following filename structure: \cr

    \emph{plotname} //"Anomaly_"// index name // plot_value // aggregation // anomyears // "_refyears-"//refyears// \emph{output} \cr e.g.: \cr
    \emph{plotname}"Anomaly_dry_days_mean_MAM_1990_refyears1981-2010.png \cr
}

\section{Details}{


When using ensemble data, the plot value is calculated over the whole ensemble (i.e. ensemble mean over the corresponding time period).
}

\examples{

data(object_st, object_grid)
autoplot_anomaly_map(
      dat_grid = object_grid, dat_p = object_st,
      anomyears=1990,
      index = "tnn", index_args = list(aggt = "seasonal", selagg = "JJA"))

data(object_hc_grid)
index_args=list(aggt = "dates", start_days = "0000-01-15", end_days = "0000-04-15")
autoplot_anomaly_map(
      dat_grid = object_hc_grid,
      anomyears=1981,refyears=1991:2010,
      index = "dd", index_args = index_args)

}
\seealso{
Other autoplot_functions: 
\code{\link{autoplot_anomaly_ts}()},
\code{\link{autoplot_climatology_map}()},
\code{\link{autoplot_forecast_map}()},
\code{\link{autoplot_forecast_spi}()},
\code{\link{autoplot_forecast_stations}()},
\code{\link{autoplot_overview_stations}()},
\code{\link{autoplot_trend_map}()},
\code{\link{autoplot_ts_stations}()},
\code{\link{autoplot_verification_map}()}
}
\concept{autoplot_functions}
