% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_spi_forecast.R
\name{autoplot_forecast_spi}
\alias{autoplot_forecast_spi}
\title{Wrapper function to calculate and plot timerseries of SPI from station data with ensemble seasonal forecast}
\usage{
autoplot_forecast_spi(
  obs_p,
  fc_p,
  index = "spi_forecast",
  index_args,
  selpoints = NULL,
  output = NULL,
  plotdir,
  plotname = "",
  plotstart,
  plot_title = TRUE,
  title = "",
  text_cex = 1,
  ...
)
}
\arguments{
\item{obs_p}{Station observations. Climindvis-object with type = "p" generated by make_object.}

\item{fc_p}{Forecasts at stations. Climindvis-object with type = "fc_p" generated by \code{\link{make_object}}.}

\item{index}{Character string. This is set to index="spi_forecast" and does not have to be provided.}

\item{index_args}{List of arguments for index. For arguments see \code{\link{index_arguments.spi_forecast}}}

\item{selpoints}{Array to select subset of points to plot. There are two options:
 \itemize{
\item Character array of station names (works only if names are defined in climindvis_index objects ...$data_info$pnames and if they are the same for all objects in climindvis_index_list)
\item Integer array of station indices (works only if  coordinates of all objects in climindvis_index_list are the same)
}}

\item{output}{Format of image to be plotted ("png", "jpeg", "tiff", "pdf","dev.new").  Default = NULL, output to null device (getOption("device") will show you the null device. If you are not using RStudio, NULL and dev.new will give the same output).}

\item{plotdir}{Character string of path of directory for saving plots. Ignored if  \emph{output} is NULL or "dev.new".}

\item{plotname}{Optional. Character string. First part of plot name followed by character string containing index name/aggregation/time period/... depending on the autoplot function. Default = "".}

\item{plotstart}{Numeric. Chosen start year of historic time series plot. For a reference period of e.g. 1981-2010 and a forecast for 2011, plotstart can be set to 2009 and only 2009-2010 and the forecast 2011 is displayed.}

\item{plot_title}{Logical. Add plot title to graphic. Default = TRUE.}

\item{title}{Optional. Character string. If provided added above the plot in addition to default information from the respective function (e.g. index name/aggregation/time period/...). Default = "", no additional title is written.}

\item{trendplots.}{No trendplots implemented in this graphic. Set to FALSE.}
}
\description{
The function returns a plot with the historical SPI values as a barplot and the ensemble forecasts as boxplots.
The function works with station data only and needs a observational time series and an ensemble forecast as input.
To see some example outputs of this function, look at the vignette "autoplot-functions" accessible through the package help main page or by typing \cr
\emph{vignette("autoplot-functions",package="ClimIndVis")} into the console.
}
\section{Details}{

For the SPI calculation, aggt needs to be set to "monthly". For further arguments see \code{\link{index_arguments.spi_forecast}} \cr

The forecast dates need to continue the time series of the historical dataset. If the observational data ends on 2010-04-30, the forecast data needs to start 2010-05-01. Thus, the SPI is calculated for the overlapping period. If there is a time gap between forecast and historical data, no overlapping period is calculated. \cr
The function can only be applied to station data and not to gridded data. Note, that you have to provide the same amount of station data in the two climdindvis-objects forecast and observation (obs_p, fc_p). No checks are executed, whether the coordinates of the observational data corresponds to the coordinates of the forecast data.

The SPI calculation is based on the SCI package from Lukas Gudmundsson & James Stagge. For the argument distribution, provide a distribution for which the corresponding density function (dname), the corresponding distribution function (pname) and the quantile function (qname) must be defined (see for example GammaDist). Distributions are listed in the package stats.\cr

For SPI with forecast no trend can be calculated at the moment.
}

\examples{

#' ## Load example climindvis objects:
data(object_st, object_fc_st)

autoplot_forecast_spi(
     obs_p = object_st,fc_p = object_fc_st, index = "spi_forecast",
     index_args = list(aggt = "monthly", timescale= 3),plotstart=2008)
}
\references{
Stagee, J.H. ; Tallaksen, L.M.; Gudmundsson, L.; van Loon, A.; Stahl, K.: Candidate Distributions for Climatological Drought Indices (SPI and SPEI), 2015, International Journal of Climatology, 35, 4027-4040, doi:10.1002/joc.4267.
}
\seealso{
Other autoplot_functions: 
\code{\link{autoplot_anomaly_map}()},
\code{\link{autoplot_anomaly_ts}()},
\code{\link{autoplot_climatology_map}()},
\code{\link{autoplot_forecast_map}()},
\code{\link{autoplot_forecast_stations}()},
\code{\link{autoplot_overview_stations}()},
\code{\link{autoplot_trend_map}()},
\code{\link{autoplot_ts_stations}()},
\code{\link{autoplot_verification_map}()}
}
\concept{autoplot_functions}
