% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_overview_stations.R
\name{autoplot_overview_stations}
\alias{autoplot_overview_stations}
\title{wrapper function to show boxplots and index values of selected years for different indices for each station.}
\usage{
autoplot_overview_stations(dat_p, indices, indices_args,
  selpoints = NULL, addyears = NULL, yearcols = 1:length(addyears),
  yearpchs = 18, yearcex = 3, NAmaxClim = 20, output = NULL,
  plotdir, plotname = "", plot_title = TRUE, title = "",
  graphic_device)
}
\arguments{
\item{dat_p}{Climindvis-object with type = "p" generated by \code{\link{make_object}}.}

\item{indices}{Character array of indices to be calculated.}

\item{indices_args}{List containing a list of index arguments for each index. See \code{\link{calc_index}}.}

\item{selpoints}{Optional integer array of stations to select for plotting. Default= NULL (all stations are plottet).}

\item{addyears}{Optional array of years (integer or character) to add as points to the plot. Default=NULL.}

\item{yearcols}{Optional array of same length as addyears with colors for marking selected years. Entries have to be valid arguments to \code{\link[grDevices]{col2rgb}}. Default=1:length(addyears)}

\item{yearpchs}{Optional array of pch for each year. Default= 18}

\item{yearcex}{Optional magnification to be used for year symbols. Default=3}

\item{NAmaxClim}{Integer [0,100]. Maximum percentage of years with missing values for calculation of climatology. If exceeded value for respective grid point/station is set to NA. Default = 20 ( = 20"\%").}

\item{output}{Format of image to be plotted ("png", "jpg", "tiff", "pdf","dev.new").  Default = NULL, output to null device (getOption("device") will show you the null device. If you are not using RStudio, NULL and dev.new will give the same output).}

\item{plotdir}{Character string of path of directory for saving plots. Ignored if  \emph{output} is NULL or "dev.new".}

\item{plotname}{Optional. Character string. First part of plot name followed by character string containing index name/aggregation/time period/... depending on the autoplot function. Default = "".}

\item{plot_title}{Logical. Add plot title to graphic. Default = TRUE.}

\item{title}{Optional. Character string. If provided added above the plot in addition to default information from the respective function (e.g. index name/aggregation/time period/...). Default = "", no additional title is written.}

\item{graphic_device}{Optional arguments for graphics devices (in addition to \emph{filename}, \emph{width} and \emph{height} ), depending on the output chosen see \code{\link[grDevices]{png}} or \code{\link[grDevices]{pdf}}. Ignored if  \emph{output} is NULL or "dev.new".}
}
\description{
This function calculates different indices at station level and then combines their climatology as boxplots in one graphic. Optionally selected years can be added as points (e.g. the current year or some years of special interest.) \cr
To see some example outputs of this function, look at the vignette "autoplot-functions" accesible through the package help main page or by typing \cr
\emph{vignette("autoplot-functions",package="ClimIndVis")} into the console.
}
\section{Output}{

This function returns one plot per station and temporal aggregation whch is covered by the data( e.g. if aggt="seasonal" it would return 4 graphics, one for each season).

If output is not NULL or "dev.new" the graphics are directly saved in the user specified directory (\emph{plotdir}) with the following filename structure: \cr

    \emph{plotname} // "overview_indices" // aggregation // years // addyears //station_name// \emph{output} \cr e.g.: \cr
    \emph{plotname}"_overview_indices_MAM_1981-2010_add2018_station2.png \cr
}

\section{Details}{

The selected indices are grouped by their units (e.g. mm, \%, no of days,...) before plotting.
}

\examples{
indices = c("dd", "fd", "th", "tx90p", "cdd", "sdii")

aggt = "seasonal"
indices_args = list(dd = list(aggt = aggt),
                   fd = list(aggt = aggt),
                   th = list(aggt = aggt, th = 5, thvar = "prec", op = ">"),
                   tx90p = list(aggt = aggt),
                   cdd = list(aggt = aggt),
                   sdii = list(aggt = aggt)
                   )


autoplot_overview_stations(dat_p = object_st, indices = indices, indices_args = indices_args,
                          addyears = 1981, selpoints = 1)

}
\seealso{
Other autoplot_functions: \code{\link{autoplot_anomaly_map}},
  \code{\link{autoplot_anomaly_ts}},
  \code{\link{autoplot_climatology_map}},
  \code{\link{autoplot_forecast_map}},
  \code{\link{autoplot_forecast_spi}},
  \code{\link{autoplot_forecast_stations}},
  \code{\link{autoplot_trend_map}},
  \code{\link{autoplot_ts_stations}},
  \code{\link{autoplot_verification_map}}
}
\concept{autoplot_functions}
