% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_trend_map.R
\name{autoplot_trend_map}
\alias{autoplot_trend_map}
\title{Wrapper function to calculate and plot trend of indices.}
\usage{
autoplot_trend_map(dat_grid, dat_p, index, index_args = list(), abs = FALSE,
  sig_lev = 0.95, NAmaxTrend = 20, selyears = NULL, title = "",
  plot_title = TRUE, plot_args = list(), output = NULL, plotdir,
  plotname = "")
}
\arguments{
\item{dat_grid}{Climindvis-object with type = "grid" generated by \code{\link{make_object}}.}

\item{dat_p}{Climindvis-object with type = "p" generated by \code{\link{make_object}}.}

\item{index}{Name of index to be calculated.}

\item{index_args}{List of arguments for index. See \code{\link{calc_index}}.}

\item{abs}{Logical. Plot absolute trend (unit/decade) instead of relative trend (percentage change relative to middle time). Default = TRUE. For indices with continuous temperature values (e.g. TNN, TNX, etc.), no relative trend is calculated.}

\item{sig_lev}{Level of significance. Only significant trend estimates are plotted. Default = 0.95.}

\item{NAmaxTrend}{Integer [0,100]. Maximum percentage of years with missing values for calculation of trend. If exceeded value for respective grid point/station is set to NA. Default = 20 ( = 20"\%").}

\item{selyears}{Integer array of years to be selected from data. For indices with quantiles, years as defined in index_args are used for quantile calculation, but only selected years are plotted. For functions using seasonal forecast data, selyears is only applied to hindcasts.}

\item{title}{Optional. Character string. If provided added above the plot in addition to default information from the respective function (e.g. index name/aggregation/time period/...). Default = "", no additional title is written.}

\item{plot_title}{Logical. Add plot title to graphic. Default = TRUE.}

\item{plot_args}{Optional. List of arguments passed to plotting function, see \code{\link{plot_args_map}}.}

\item{output}{Format of image to be plotted ("png", "jpg", "tiff", "pdf","dev.new").  Default = NULL, output to null.device}

\item{plotdir}{Character string of path of directory for saving plots. Ignored if  \emph{output} is NULL or "dev.new".}

\item{plotname}{Optional. Character string. First part of plot name followed by character string containing index name/aggregation/time period/... depending on the autoplot function. Default = "".}
}
\description{
This function calls the function \code{\link{calc_index}} and plots the significant trend values in a map. It can be used for either gridded or station data or a combination of both.\cr
To see some example outputs of this function, look at the vignette "autoplot-functions" accessible through the package help main page or by typing  \cr
\emph{vignette("autoplot-functions",package="ClimIndVis")} into the console.
}
\section{Details}{

For every index a default method is used for the trend calculation. The default trend estimation and test depends on the nature of the calculated index: \cr

For continuous data (e.g. sum, mean, prcptot, spi) by default an ordinary linear regression is calculated with time as predictor for the trend estimate and applying a students-t test for significance testing. If the data is not normaly distributed, a logarithmic transformation is applied before the trend calculation, e.g. for precipitation sums. \cr

For count data (e.g dd, cdd, cwd) by default a logistic regression is calculated with time as predictor for the trend estimation and applying a students-t test for significance testing. In the calculation, a correction for overdispersion is applied. \cr

If you choose trend = "MannKendall" instead of trend = TRUE the non-parametric Mann-Kendall test based on the relative ranking in the series is applied and a Theil-Sen slope is estimated. See e.g. Yue et al. 2002 or Mann (1945), Kendall (1975) for further references.
}

\section{Output}{

This function returns one plot per temporal aggregation which is covered by the data( e.g. aggt="seasonal" returns 4 graphics, one for each season).

If output is not NULL or "X11" the graphics are saved in the user specified directory (\emph{plotdir}) with the following filename structure: \cr

    \emph{plotname} // index name // aggregation // "abs-/rel_trend" // years // \emph{output} \cr e.g.: \cr
    \emph{plotname}"_dry_days_MAM_rel-trend_1981-2010.png \cr
}

\examples{
data(object_st,object_grid)
autoplot_trend_map(dat_grid=object_grid,
                         index="mean",index_args=list(var="tmax",aggt="seasonal",selagg="SON"))

}
\seealso{
Other autoplot_functions: \code{\link{autoplot_climatology_map}},
  \code{\link{autoplot_forecast_map}},
  \code{\link{autoplot_forecast_spi}},
  \code{\link{autoplot_forecast_stations}},
  \code{\link{autoplot_overview_stations}},
  \code{\link{autoplot_ts_stations}},
  \code{\link{autoplot_verification_map}}
}
