% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fc_cats.R
\name{calc_fc_cats}
\alias{calc_fc_cats}
\title{Get probabilities of forecast categories}
\usage{
calc_fc_cats(fc, hc, index, index_args = list(), selyears = NULL, prob,
  ncat = 3, counts = FALSE)
}
\arguments{
\item{fc}{Forecasts. "climindvis_index" object  of type fc_p/fc_grid as output from \code{\link{make_object}} using an ensemble daily seasonal forecasts.}

\item{hc}{Hindcasts. "climindvis_index" object  of type hc_p/hc_grid as output from \code{\link{make_object}} using an ensemble of  daily seasonal hindcasts.}

\item{index}{Name of index to be calculated.}

\item{index_args}{List of arguments for index. See \code{\link{calc_index}}.}

\item{selyears}{Integer array of years to be selected from data. For indices with quantiles, years as defined in index_args are used for quantile calculation, but only selected years are plotted. For functions using seasonal forecast data, selyears is only applied to hindcasts.}

\item{prob}{Probability thresholds for converting continuous forecasts to category forecasts. If not provided then ncat equidistant categories are calulated: probs= 1:(ncat-1)/ncat.}

\item{ncat}{Number of forecast categories. Default = 3.}

\item{counts}{Logical. Should number of ensemble members per category be returned instead of probabilities. Default=FALSE}
}
\value{
Object of class \emph{index} and "climindvis_fc" with following entries:
\itemize{
  \item cprob/ccount: Array of forecast probabilities/counts for each category with named dimensions
  \item {cpinfo:} {Named list with additional information. Absolute category threshold values from hindcasts (\emph{prob_th}) and the corresponding probability tresholds (\emph{prob}), number of cateogories (\emph{ncat}) and dimensions and dimension names of cprob (\emph{cpdims},\emph{cpdimnames}).}
  \item lon: Array of longitudes
  \item lat: Array of latitudes
  \item index_info: Named list with information on index such as index name and aggregation type
  \item {data_info$fc and $hc :}{Data_info taken from fc and hc input objects, e.g. name of dataset}
}
}
\description{
Calculates index for hindcasts and forecasts and then derives category thresholds from hindcasts and calculates forecast probabilities for each category.
The function works with gridded or station data (all input data needs to be of same type) and first calculates the chosen index (using \code{\link{calc_index}}) for hindcasts and forecasts and then divides the forecast into categories (e.g. terciles) based on the climatology of the hindcasts. It outputs the forecast probabilities for each category as well as the category thresholds.
}
\section{Note}{

This function either returns the forecast probabilities (\emph{cprob}) or, if \emph{counts} is set to TRUE, the number of ensemble members per category (\emph{ccount}) . Forecast probabilities are calculated as follows:  \emph{ccounts}/number of ensemble members *100.
These values are rounded to the nearest integer, therefore adding the forecast probabilities of all categories may not always add up to 100.
Note that for some forecast the year dimension might be greater than one, e.g. for a November forecast and monthly temporal aggregation, the first two months (Nov. + Dez) are from another year than the following months.
}

\examples{

## Load example climindvis objects:

data("object_hc_st", "object_fc_st")

calc_fc_cats(fc = object_fc_st, hc = object_hc_st, index = "tnn", index_args = list(aggt="seasonal"))


data("object_hc_grid", "object_fc_grid")

calc_fc_cats(fc = object_fc_grid, hc = object_hc_grid, index = "tnn", index_args = list(aggt="seasonal"))


}
