% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_trends.R
\name{calc_index_trend}
\alias{calc_index_trend}
\title{A function to calculate a logistic, linear or non-parametric trend of an index.}
\usage{
calc_index_trend(index, trend, targs, time = NULL)
}
\arguments{
\item{index}{Array (dim>1) of index or variable for which trend is calculated.}

\item{trend}{Logical or character. If trend is a character string with "MannKendall" a non-parametric Mann-Kendall test is applied and a TeilSen slope estimated.}

\item{targs}{List of arguments for calculation of trends. Depending on index or variable different methods for the trend calculations have to be chosen.
\itemize{
  \item \strong{method} Three methods are available: (1) logit ("log_reg") or (2) linear regression ("lin_reg") or (3) "MannKendall" calculates a TheilSen slope and a MannKendall test.
  \item \strong{count} set to TRUE, if the data is count data.
  \item \strong{log_trans} if the data is not normaly distributed, set log_trans to TRUE and a logarithmic transformation is applied.
  \item \strong{NAmaxTrend} Maximum value of NAs for calculation of trend. If value is exceeded, trend is set to NA.

      }}

\item{time}{Vector of years to base trend calculation on. If not provided and data is not a named vector or array, calculation is aborted.}
}
\value{
The function returns a list with two elements "statistics" and "data". Data is an array with the variable dimension and four additional dimensions for fitted trend line, upper and lower confidence interval and a local polynomial regression fitting. "statistics" is an array with the variable dimensions and additional a dimension for start year, end year, p-value, relative trend, absolute trend and trend method.
}
\description{
A function to calculate a logistic, linear or non-parametric trend of an index.
}
\section{Details}{
 The default trend estimation and test depends on the nature of the calculated index: \cr

For continuous data (e.g. sum, mean, prcptot, spi) ny default an ordinary linear regression is calculated with time as predictor for the trend estimate and a students-t test. If the data is not normaly distributed, a logarithmic transformation is applied before the trend calculation, e.g. for . \cr

For count data (e.g dd, cdd, cwd) by default a logistic regression is calculated with time as predictor for the trend estimation and a students-t test. In the calculation a correction for overdispersion is applied. \cr

If you choose trend = "MannKendall" instead of trend = TRUE the non-parametric Mann-Kendall test based on the relative ranking in the series is applied and a Theil-Sen slope is estimated. See e.g. Yue et al. 2002 or Mann (1945), Kendall (1975).
}

\keyword{internal}
