% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_main.R
\name{calculate_index.climindvis}
\alias{calculate_index.climindvis}
\title{default index function for climindvis objects}
\usage{
\method{calculate_index}{climindvis}(climindvis, aggt, ...)
}
\arguments{
\item{climindvis}{climindvis object generated using the function make_object}

\item{args}{List of arguments for calculation of index.  Arguments depend on type of index, for more information see documentation of calculate_index}
}
\description{
default index function for climindvis objects
}
\seealso{
\link{make_object}

\link{calc_index}

Other calculate_index: \code{\link{calculate_index.spi_forecast}}
}
\concept{calculate_index}
\keyword{internal}
