% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_index_spei.R
\name{calculate_index.spi_forecast}
\alias{calculate_index.spi_forecast}
\title{default index function for climindvis objects}
\usage{
\method{calculate_index}{spi_forecast}(climindvis, aggt, ...)
}
\arguments{
\item{climindvis}{Climindvis object generated using the function make_object.}

\item{forecast}{Climindvis object with forecast data generated using the function make_object}

\item{args}{List of arguments for calculation of index.  Arguments depend on type of index, for more information see documentation of calculate_index}
}
\value{
The function returns a climindvis_index_object with an additonal entry <index_fcst>. This
}
\description{
default index function for climindvis objects
}
\section{Details}{

The SPI calculation is based on the SCI package from Lukas Gudmundsson & James Stagge. For integrating forecasts into the SPI plot, the timeseries of forecast and historical data are checked for the overlapping period. For this period a mean
}

\references{
Stagee, J.H. ; Tallaksen, L.M.; Gudmundsson, L.; van Loon, A.; Stahl, K.: Candidate Distributions for Climatological Drought Indices (SPI and SPEI), 2015, International Journal of Climatology, 35, 4027-4040, doi:10.1002/joc.4267.
}
\seealso{
\link{make_object}

\code{\link{calc_index}}

Other calculate_index: \code{\link{calculate_index.climindvis}}
}
\keyword{internal}
