% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_get_arguments.R
\name{index_arguments.qrange}
\alias{index_arguments.qrange}
\title{arguments for index qrange}
\usage{
\method{index_arguments}{qrange}(
  climindvis,
  NAmaxAgg = 20,
  trend = FALSE,
  NAmaxTrend = 20,
  percentile1,
  percentile2,
  var,
  ...
)
}
\arguments{
\item{NAmaxAgg}{Maximum number of missing values in daily data of each aggregation period (e.g. in all January values for the whole time series). Value between 0 (=0\%; no missing values allowed) and 100 (=100\%; all values can be missing values). If the number of missing values is higher than NAmaxAgg the index value for the respective grid point/station is set to NA. Default = 20 (20\%).}

\item{trend}{Logial (or character). Calculates trends with a predefined method for the chosen index. For information on default method, see  \code{\link{trend_methods}}. You can chose to calculate a non-parameteric trend, by setting trend to a character string "MannKendall". This calculates a TheilSen slope and MannKendall test. Default = FALSE.}

\item{NAmaxTrend}{Maximum number of years in time series of index which are missing values. Value between 0 (=0\%); no years with missing values allowed) and 100 (=100\%); all years can be missing values). If exceeded, the trend is set to NA for the respective grid point/station. Note that the value is relative to the length of the time series. Default = 20 (20\%).}
}
\description{
arguments for index qrange
}
\section{DETAILS}{

This index calculated the difference between the two percentile values for each aggregation and year seperately, e.g. in the case of the 25th and the 75th percentile the interquantile range is calculated.
}

\examples{
data(object_st) # load example objects
calc_index(object_st,index="qrange",aggt="seasonal",precentile1=25,percentile2=75,var="tmax")
}
\keyword{internal}
