% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_get_arguments.R
\name{index_arguments.rainy_season_dur}
\alias{index_arguments.rainy_season_dur}
\alias{index_arguments.rainy_season_dur}
\title{arguments for index rainy_season_dur}
\usage{
\method{index_arguments}{rainy_season_dur}(climindvis, rs_method, days,
  dd_th = 1, nval = NA, trend = FALSE, NAmaxTrend = 20, NAdays = 0,
  ...)

\method{index_arguments}{rainy_season_dur}(climindvis, rs_method, days,
  dd_th = 1, nval = NA, trend = FALSE, NAmaxTrend = 20, NAdays = 0,
  ...)
}
\arguments{
\item{climindvis}{Object of type climindvis as returned by \code{\link{make_object}} which must contain the variable precipitation (prec)}

\item{dd_th}{Dry day threshold}

\item{trend}{Logial (or character). Calculates trends with a predefined method for the chosen index. For information on default method, see  \code{\link{trend_methods}}. You can chose to calculate a non-parameteric trend, by setting trend to a character string "MannKendall". This calculates a TheilSen slope and MannKendall test. Default = FALSE.}

\item{NAmaxTrend}{Maximum number of years in time series of index which are missing values. Value between 0 (=0\%); no years with missing values allowed) and 100 (=100\%); all years can be missing values). If exceeded, the trend is set to NA for the respective grid point/station. Note that the value is relative to the length of the time series. Default = 20 (20\%).}

\item{NAdays}{Number of NA days allowed per decade for index <<sos>>. Default=0.}

\item{rs_method:}{method for calculation of index (further indicator arguments for each rs_method are given in brackets and defined below, arguments without default that have to be provided are in bold font ):
\itemize{
  \item garcia (nval): sum(3days) > 20mm and maximum number of consecutive dry days within the next 30 days <=10 (definition from Garcia et al.,2007 )

  \item gurgiser (nval): sum(7days) > 10mm and number of consecutive wet days   within the next 30 days > 10 (definition from Gurgiser et al.,2017)
  }}

\item{th}{Threshold for precipitation sum of <<days>> consecutive days}

\item{rs_method:}{method for calculation of index (further indicator arguments for each rs_method are given in brackets and defined below, arguments without default that have to be provided are in bold font ):
\itemize{
  \item garcia (nval): sum(3days) > 20mm and maximum number of consecutive dry days within the next 30 days <=10 (definition from Garcia et al.,2007 )

  \item gurgiser (nval): sum(7days) > 10mm and number of consecutive wet days   within the next 30 days > 10 (definition from Gurgiser et al.,2017)
  }}

\item{th}{Threshold for precipitation sum of <<days>> consecutive days}

\item{dd_th}{Dry day threshold}

\item{NAdays}{Number of NA days allowed per decade for index <<sos>>. Default=0.}
}
\description{
arguments for index rainy_season_dur

arguments for index rainy_season_dur
}
\section{DETAILS}{

The duration of the rainy season is calculated using the same functions for the start and end of the rainy season. The arguments for the aggregations have to be provided differently. <<day start>> indicates the first day of the calculation for the start of the rainy season and <<day end>> the first day of the time series for calculating the end of the rainy season. nstart/nend are the number of days after day_start/day_end to consider for the calculation (defaults to 300days). If the criterion for start or end are not met the value of the respective year is set to NA.


xxx
}

\examples{
data(object_st) # load example objects
calc_index(object_st,index="rainy_season_start",aggt="dates",start_days="0000-08-01",end_days="0000-03-01",rs_method="gurgiser")

data(object_st) # load example objects
calc_index(object_st,index="rainy_season_start",aggt="dates",start_days="0000-08-01",end_days="0000-03-01",rs_method="gurgiser")

}
\keyword{internal}
