% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_get_arguments.R
\name{index_arguments}
\alias{index_arguments}
\title{get arguments index
Get arguments for index to pass to function and check if variable from which index is calculated exists
Generic function for getting climindvis index arguments}
\usage{
index_arguments(climindvis, ...)
}
\arguments{
\item{NAmaxAgg}{Maximum number of missing values in daily data of each aggregation period (e.g. in all January values for the whole time series). Value between 0 (=0\%; no missing values allowed) and 100 (=100\%; all values can be missing values). If the number of missing values is higher than NAmaxAgg the index value for the respective grid point/station is set to NA. Default = 20 (20\%).}
}
\description{
get arguments index
Get arguments for index to pass to function and check if variable from which index is calculated exists
Generic function for getting climindvis index arguments
}
\keyword{internal}
