% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_get_arguments.R
\name{index_arguments.spi_forecast}
\alias{index_arguments.spi_forecast}
\title{arguments for index spi_forecast}
\usage{
\method{index_arguments}{spi_forecast}(climindvis, timescale = 6,
  forecast = TRUE, ref = NULL, distribution = "gamma", trend = FALSE,
  trendargs, NAmaxAgg = 20, ...)
}
\arguments{
\item{timescale}{Integer. Timescale for monthly SPI calculation, i.e.running mean for 1-monthly,3-monthly, etc. period. Default = 6.}

\item{forecast}{Climindvis object with forecast data.}

\item{ref}{Vector with start and end year for reference period. If no reference period is specified, the full period is used as reference period.}

\item{distribution}{A character string naming a distribution of package stats (see for example GammaDist). For more information on the choice of the distribution, see section Details. Default = "gamma".}

\item{trend}{Logial (or character). Calculates trends with a predefined method for the chosen index. For information on default method, see  \code{\link{calc_index_trend}}. You can chose to calculate a non-parameteric trend, by setting trend to a character string "MannKendall". This calculates a TheilSen slope and MannKendall test. Default = TRUE.}

\item{NAmaxAgg}{Maximum number of missing values in daily data of each aggregation period (e.g. in all January values for the whole time series). Value between 0 (=0\%; no missing values allowed) and 100 (=100\%; all values can be missing values). If the number of missing values is higher than NAmaxAgg the index value for the respective grid point/station is set to NA. Default = 20 (20\%).}
}
\description{
arguments for index spi_forecast
}
\section{Details}{

The SPI calculation is based on the SCI package from Lukas Gudmundsson & James Stagge. For integrating forecasts into the SPI plot, the timeseries of forecast and historical data are checked for the overlapping period. For this period a mean
}

\references{
Stagee, J.H. ; Tallaksen, L.M.; Gudmundsson, L.; van Loon, A.; Stahl, K.: Candidate Distributions for Climatological Drought Indices (SPI and SPEI), 2015, International Journal of Climatology, 35, 4027-4040, doi:10.1002/joc.4267.
}
\keyword{internal}
