% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_get_arguments.R
\name{index_arguments.sud}
\alias{index_arguments.sud}
\title{arguments for index sud}
\usage{
\method{index_arguments}{sud}(climindvis, iformat = "perc", NAmaxAgg = 20,
  trend = FALSE, NAmaxTrend = 20, dd_threshold = 1, ...)
}
\arguments{
\item{climindvis}{Object of type climindvis as returned by \code{\link{make_object}} which must contain the variable maximum temperature (tmax)}

\item{iformat}{Format for indicator "perc"(default)/"days"}

\item{NAmaxAgg}{Maximum number of missing values in daily data of each aggregation period (e.g. in all January values for the whole time series). Value between 0 (=0\%; no missing values allowed) and 100 (=100\%; all values can be missing values). If the number of missing values is higher than NAmaxAgg the index value for the respective grid point/station is set to NA. Default = 20 (20\%).}

\item{trend}{Logial (or character). Calculates trends with a predefined method for the chosen index. For information on default method, see  \code{\link{trend_methods}}. You can chose to calculate a non-parameteric trend, by setting trend to a character string "MannKendall". This calculates a TheilSen slope and MannKendall test. Default = FALSE.}

\item{NAmaxTrend}{Maximum number of years in time series of index which are missing values. Value between 0 (=0\%); no years with missing values allowed) and 100 (=100\%); all years can be missing values). If exceeded, the trend is set to NA for the respective grid point/station. Note that the value is relative to the length of the time series. Default = 20 (20\%).}

\item{dd_threshold}{Precipitation threshold for dry days in mm. Default = 1}
}
\description{
arguments for index sud
}
\examples{
data(object_st) # load example objects
calc_index(object_st,index="dd",aggt="monthly")
}
\keyword{internal}
