% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_object.R
\encoding{UTF-8}
\name{make_object}
\alias{make_object}
\title{Function to generate a climindvis object}
\usage{
make_object(tmin = NULL, tmax = NULL, prec = NULL, tavg = NULL,
  dates_tmin = NULL, dates_tmax = NULL, dates_prec = NULL,
  dates_tavg = NULL, lon, lat, data_range = NULL, data_info)
}
\arguments{
\item{tmin}{Daily minimum temperature data in degrees C (see NOTES)}

\item{tmax}{Daily maximum temperature data in  degrees C (see NOTES)}

\item{prec}{Daily total precipitation data in mm (see NOTES)}

\item{tavg}{Daily average temperature data in degrees C (see NOTES)}

\item{dates_tmin}{Dates for the daily minimum temperature data (see NOTES)}

\item{dates_tmax}{Dates for the daily maximum temperature data (see NOTES)}

\item{dates_prec}{Dates for the daily precipitation data (see NOTES)}

\item{dates_tavg}{Dates for the daily average temperature data (see NOTES)}

\item{lon}{Longitudes of input data (1 or 2dim array)}

\item{lat}{Latitudes of input data (1 or 2dim array)}

\item{data_range}{Vector with start and end year of calculation period (optional). If supplied, all data is cut to this period. If missing, a period from the earliest until the last date value is defined.}

\item{data_info}{List of data-specific information with the following elements:
\itemize{
 \item type [character]: data type (see NOTES)
    \itemize{
      \item p,p_hc,p_fc: point data for observations/reanalyses, hindcasts,forecasts. For a single station, see NOTES.
      \item grid,grid_hc,grid_fc: gridded data for observations/reanalysis, hindcasts,forecasts
    }
 \item date_format  [character]:
  \itemize{
      \item t1d: In this case data has the dimension spatial-dims ( x ensemble dims) x time and time is a 1d array
      \item t2d: In this case data has the dimension spatial-dims ( x ensemble dims) x no days x no years and time is a named list of length years with each list element containing the dates for one year, the name of the list element should correspond to the year of the first list entry. In the case of a turn of the year, the name in time list corresponds to the year of the first entry of the time array. NOTE2: in case of forecast data, the array should have the last dimension as year dimension of length one (if the year dimension is missing it is set to one) and years can be given as list or as array of dates. NOTE3: In this case, all variables need to have the same time steps, i.e. dates_xxx need to be equal for all variables.
    }
 \item data_name: Name of dataset (e.g. "Station data" or "ECMWF S4 hindcasts"). Nessesary for plot functions

 \item pnames: Character array of station names (optional). If provided they can be used for plotting. For point data (type="p*") only.
}}
}
\value{
The function returns a climindvis object with the following entries:
\itemize{
\item data: Named list containing an array of data (tmin/tmax/prec/tavg) of dimension: spatial dimensions x (ensemble dimensions) x no of timesteps
\item time: Array of timesteps which are common for all input data.
\item lon: Array of longitudes common for all input data.
\item lat: Array of latitudes common for all input data.
\item data_info: data_info as assed to function.
\item time_factors: time factors used for calculation of temporal aggregation of indices.
}
}
\description{
This function creates a climindvis object needed as input for all other functions of the package. The function performs several checks on the data and formats the data as it is needed by the other functions. For the different data types, see the parameter description below.
}
\section{NOTES}{


The data input is an array where the dimensions depend on the type of data. Generally the data must be structured in the following way:
spatial dimensions,ensemble dimension (in case of hindcasts or forecasts), time dimensions. In case of single stations, it must still be a matrix with two dimensions; one for the station and one for the time.
At least one of the 4 variables (and the respective dates) must be provided, otherwise the function gives an error.

As mentioned in date_format, the format of dates_* can either be an 1dim array of all the dates or in the case of time slice data, a list of the dates for each year.

For example data and example objects see \code{\link{example_data}} \code{\link{example_climindvis_objects}}.
}

\examples{
# make climindvis object with station data

data(data_st)
climindvis_st <- make_object(tmin = data_st$tmin, tmax = data_st$tmax, tavg = data_st$tvag, prec = data_st$prec,
dates_tmin = data_st$time, dates_tmax = data_st$time, dates_tavg = data_st$time, dates_prec = data_st$time,
lon = data_st$lon, lat = data_st$lat,  data_info = data_st$data_info, data_range=c(1950,2000))
class(climindvis_st)

# make climindvis object with gridded data

data(data_grid)
climindvis_grid <- make_object(
    tmin = data_grid$tmin,tmax = data_grid$tmax, prec = data_grid$prec,
    dates_tmin = data_grid$time, dates_tmax = data_grid$time, dates_prec = data_grid$time,
    lon = data_grid$lon, lat = data_grid$lat,  data_info = data_grid$data_info)
class(climindvis_grid)

# make climindvis object with gridded hindcast data

data(data_hc_grid)
climindvis_grid_hc <- make_object(
    tmin = data_hc_grid$tmin, tmax = data_hc_grid$tmax, prec = data_hc_grid$prec,
    dates_tmin = data_hc_grid$time, dates_tmax = data_hc_grid$time, dates_prec = data_hc_grid$time,
    lon = data_hc_grid$lon, lat = data_hc_grid$lat,  data_info = data_hc_grid$data_info)
class(climindvis_grid_hc)
}
