% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_documentation.R
\name{plot_doc}
\alias{plot_doc}
\title{plot docu}
\arguments{
\item{obs}{Observations. "climindvis" object of type p/grid as output from \code{\link{make_object}} by using daily observational or Re-analysis data.}

\item{hc}{Hindcasts. "climindvis_index" object  of type hc_p/hc_grid as output from \code{\link{make_object}} using an ensemble of  daily seasonal hindcasts.}

\item{fc}{Forecasts. "climindvis_index" object  of type fc_p/fc_grid as output from \code{\link{make_object}} using an ensemble daily seasonal forecasts.}

\item{dat_p}{Climindvis-object with type = "p" generated by \code{\link{make_object}}.}

\item{dat_grid}{Climindvis-object with type = "grid" generated by \code{\link{make_object}}.}

\item{hc_p}{Hindcasts at stations. Climindvis-object with type = "hc_p" generated by \code{\link{make_object}}.}

\item{hc_grid}{Gridded hindcasts. Climindvis-object with type = "hc_grid" generated by \code{\link{make_object}}.}

\item{hc_ref_p}{Optional. Reference hindcasts at stations for calculation of skill scores. If not provided (NULL), climatology is taken as reference. Default = NULL.}

\item{hc_ref_grid}{Optional. Reference gridded hindcasts for calculation of skill scores. If not provided (NULL), climatology is taken as reference. Default = NULL.}

\item{fc_p}{Forecasts at stations. Climindvis-object with type = "fc_p" generated by \code{\link{make_object}}.}

\item{fc_grid}{Gridded forecasts. Climindvis-object with type = "fc_grid" generated by \code{\link{make_object}}.}

\item{obs_p}{Station observations. Climindvis-object with type = "p" generated by \code{\link{make_object}}. For autoplot_forecast_map and autoplot_forecast_point only needed if skill criterion is applied.}

\item{obs_grid}{Gridded observation data. Climindvis-object with type = "grid" generated by \code{\link{make_object}}. For autoplot_forecast_map and autoplot_forecast_point only needed if skill criterion is applied.}

\item{datlist_p}{Single object or list of climindvis-objects with type = "p"/"hc_p" generated by \code{\link{make_object}}.}

\item{index}{Name of index to be calculated., e.g. index = "dd" or index = "tnn". For list of indices see \code{\link{indices_list}}}

\item{index_args}{List of arguments for index. See \code{\link{indices_list}}.}

\item{indices}{Character array of indices to be calculated.}

\item{indices_args}{List containing a list of index arguments for each index. See \code{\link{calc_index}}.}

\item{selyears}{Integer array of years to be selected from data. For indices with quantiles, years as defined in index_args are used for quantile calculation, but only selected years are plotted. For functions using seasonal forecast data, selyears is only applied to hindcasts.}

\item{selpoints}{Array to select subset of points to plot. There are two options:
 \itemize{
\item Character array of station names (works only if names are defined in climindvis_index objects ...$data_info$pnames and if they are the same for all objects in climindvis_index_list)
\item Integer array of station indices (works only if  coordinates of all objects in climindvis_index_list are the same)
}}

\item{trendplots}{Logical. Plot trend line and 95\% confidence interval for index. Default = FALSE.}

\item{NAmaxClim}{Integer [0,100]. Maximum percentage of years with missing values for calculation of climatology. If exceeded value for respective grid point/station is set to NA. Default = 20 ( = 20"\%").}

\item{NAmaxTrend}{Integer [0,100]. Maximum percentage of years with missing values for calculation of trend. If exceeded value for respective grid point/station is set to NA. Default = 20 ( = 20"\%").}

\item{text_noskill}{Character string. Text to show if forecast skill is lower than noskill. Default = "no skill".}

\item{text_novar}{Character string. Text if climatology shows no or little variation and plotting of forecast does not make sense. Default = "no or little variation in climatology".}

\item{text_nocat}{Character string. Text if forecast shows no dominant forecast category (=probabilities of two or more cateogries are equal). Default = "no dominant category".}

\item{veri_metrics}{Character array of skill metrics to calculate. For metrics see \code{\link{verify_index}}.}

\item{veri_args}{Optional list of additional arguments for verification, see \code{\link{veri_args}}}

\item{ncat}{Number of forecast categories. Default = 3.}

\item{plot_climatology}{Logical. Should climatology of index be plottet as time series? If TRUE, time series with shading indicating corresponding climatological values for forecast categories are plottet. Default = FALSE.}

\item{tacholabs}{Labels for categories of the tacho plot.}

\item{output}{Format of image to be plotted ("png", "jpeg", "tiff", "pdf","dev.new").  Default = NULL, output to null device (getOption("device") will show you the null device. If you are not using RStudio, NULL and dev.new will give the same output).}

\item{plotdir}{Character string of path of directory for saving plots. Ignored if  \emph{output} is NULL or "dev.new".}

\item{plot_title}{Logical. Add plot title to graphic. Default = TRUE.}

\item{plotname}{Optional. Character string. First part of plot name followed by character string containing index name/aggregation/time period/... depending on the autoplot function. Default = "".}

\item{graphic_device}{Optional arguments for graphics devices (in addition to \emph{filename}, \emph{width} and \emph{height} ), depending on the output chosen see \code{\link[grDevices]{png}} or \code{\link[grDevices]{pdf}}. Ignored if  \emph{output} is NULL or "dev.new".}

\item{title}{Optional. Character string. If provided added above the plot in addition to default information from the respective function (e.g. index name/aggregation/time period/...). Default = "", no additional title is written.}

\item{plot_args}{Optional. List of arguments passed to plotting function, see \code{\link{plot_args_map}}.}
}
\description{
plot docu
}
\keyword{internal}
