% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dat_in_map.R
\name{plot_map_grid_points}
\alias{plot_map_grid_points}
\title{function to plot data points on map of gridded data}
\usage{
plot_map_grid_points(g_dat = NULL, g_lon = NULL, g_lat = NULL,
  g_col = rainbow(10), g_col_center = FALSE, g_breaks, g_nlev = 10,
  p_lon = NULL, p_lat = NULL, p_col_info = "same",
  p_col = rainbow(10), p_col_center = FALSE, p_dat, p_legend = 0,
  legend_center = FALSE, xlims, ylims, zlims = NULL, p_zlims,
  mask = NULL, mask_NA = NULL, ratio = 1, p_nlev = 10,
  p_breaks = NA, p_text, p_pch = 21, p_cex = 1, p_pch_s = NA,
  p_col_sig, output = NULL, outfile, plot_scale = TRUE, pwidth = 5,
  graphic_device, plot_title, lwidth = 1, plwidth = 1, trend = FALSE,
  topo = NULL, text_cex = 1, mask_col = "lightgrey",
  NA_col = "white", fc_plot = FALSE, outliers = c(FALSE, FALSE),
  units = "", ...)
}
\arguments{
\item{g_lon, g_lat:}{arrays of longitudes and latidudes  of dimension nx,ny ( or matrix of dim nx x ny if grid is not regular).}

\item{g_breaks}{(g_nlev): breaks for plotting gridded data. If not provided, g_nlev needs to be specified to calculate breaks. If g_nlev is provided, the pretty() function is used to define the breakpoints and therefore the number of breaks might differ from the g_nlev.}

\item{p_lon, p_lat:}{array of longitudes and latitudes of points of dimension np=number of points}

\item{xlims, ylims:}{x and y limits of the plot. If not specified taken from g_lon and g_lat}

\item{p_breaks}{(p_nlev): breaks for plotting point data. If p_breaks is not provided, p_nlev needs to be specified to calculate breaks. If p_col_info = "same" or "cpoints" values are taken from g_breaks (g_nlev). If p_nlev is provided, the pretty() function is used to define the breakpoints and therefore the number of breaks might differ from the p_nlev.}

\item{pwidth}{width of plotting region of output (as input for jpeg, png, ...). height is calculated in functions based on plot layout}

\item{graphic_device}{arguments for graphics devices (in addition to <<filename>>, <<width>> and <<height>> ), depending on the output chosen see \code{\link[grDevices]{png}} or \code{\link[grDevices]{pdf}}.}

\item{plot_title}{(optional), character string for plot title}

\item{lwidth}{width of legend}

\item{plwidth}{width of legend for points}

\item{topo}{Information for contour lines added to the plot. Named list with following three elements: \emph{alt}=array of altitudes of dimension nlonXnlat, \emph{lon}: array of longitudes, \emph{lat}: array of latitudes. Default=NULL.}

\item{fc_plot}{-> forecast plot, shows probabilities for all ncat forecast categories}

\item{g_dat:}{matrix of gridded data of dimension nx x ny}

\item{g_col:}{color scale for plotting gridded data as character of function}

\item{p_dat:}{array of values for data for points of dimension p (needed for p_col_info = "same" or "cbar")}

\item{p_col_info:}{type of coloring for points
\itemize{
 \item "same": same colorbar as for gridded data, p_col and p_nlev do not need to be specified and no extra legend is plotted for points but p_dat is needed
 \item "cbar": colorbar specified in p_col but different from g_col. Either predefined colors bars such as these: link to grDevices::Palettes and gplots::colorpanel and RColorBrewer or given in "" (e.g. cbar="gplots::bluered")
 \item "cpoints": discrete colors specified for each point as array in p_col (p_dat is not needed in this case)
 }}

\item{p_col:}{colors for points, see p_col_info}

\item{p_pch:}{pch of the points (see par). Values between 21:25 are implemented, default is 21.}

\item{p_cex:}{cex of points (number or array), default is 2}

\item{p_legend:}{legend to be drawn for point data. Default is no legend. Legend is not drawn if p_col_info="same".
\itemize{
 \item "cbar": colorbar, additional argumten p_nlev or p_breaks is needed
 \item "cpoints": legend of points of unique colors of p_col, p_legend_text should be specified
 }}

\item{p_pch_s:}{x and y limits of the plot. If not specified taken from g_lon and g_lat}

\item{mask:}{plot land-sea mask of data}

\item{mask_col:}{color of mask values}

\item{leg_labs:}{legend labs (for fc_plot=TRUE)}

\item{fc_catnames:}{title for categories (for fc_plot=TRUE)}

\item{leg_title}{title of legend}

\item{output:}{needs to be specified if file output is wanted, currently only "png" is implemented (default=NA)}

\item{outfile:}{character, outputfile (only if output is not NULL). final file will be of type outfile.output}
}
\description{
function to plot data points on map of gridded data
}
\keyword{internal}
