% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spi_forecast.R
\name{plot_spi_forecast}
\alias{plot_spi_forecast}
\title{function for plotting spi for station data}
\usage{
plot_spi_forecast(ind_dat, full_dat_NA, selpoints, pdims, plot_title = TRUE,
  title = "", output = NULL, plotdir = NULL, plotname = NULL,
  ylims = NULL, plotstart = NULL, pwidth = 2000, pheight = 1300,
  pres = 250, cex = 1, text_cex = 1, show.grid.h = TRUE, ...)
}
\arguments{
\item{ind_dat}{1 List of climindvis_index objects (as returned by \code{\link{calc_index}}).
You can either plot
 \itemize{
  \item different indices calculated from the same dataset (...$data_info$data_name has to be the same for all element of ind_dat) or
  \item one indice for different datasets (class of list elements and ...$index_info$iname have to be the same for all element of ind_dat)
  }}

\item{selpoints}{Array to select subset of points to plot,there are two options:
 \itemize{
\item character array of station names (works only if names are defined in climindvis_index objects ...$data_info$pnames and if they are the same for all objects in ind_dat)
\item integer array of station indices (works only if  coordinates of all objects in ind_dat are the same)
}}

\item{ylims}{Limits of xaxis for plot of format ylims=c(ymin,ymax). If not provided limits are taken from the data}

\item{plotstart}{Start year of plot, e.g plotstart = 1990 ends at end of forecast time.}

\item{pwidth, pheight, pres}{Width,Height and Resolution of plot (see width in  \code{\link[grDevices]{png}}).If output=="pdf" the heights and widths are adjusted to pwidth/res and pheight/res. Default values are pwidth=2000,pheight=1300,pres=250.}

\item{cex}{cex of plots (see par()). Default=1}

\item{output:}{type of output xxx auch write moeglich}

\item{plotdir:}{directory where files/plots are saved to.}

\item{plotname:}{string for outputnames, the final name will be outfilestring_month/season/etc..}
}
\description{
function for plotting spi for station data
}
