% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_get_arguments.R
\name{quantile_doc}
\alias{quantile_doc}
\title{quantile}
\arguments{
\item{baseperiod}{Vector of start and end year for the calculation of quantiles. If not provided, whole range of years in dataset will be used for the calculation of quantiles.}

\item{q}{Vector of quantiles, default to seq(0,1,by=0.01)}

\item{th_object}{Climindvis_index object, default=NULL. if calculating a quantile index for forecast data, a climindvis_index object of hindcasts data for the same spatial dimensions and forecast months has to be provided. Threshold values will then be taken from the hindcast object (object$index_info$quantiles). In this case all other arguments (q, baseperiod, threhold,..) are ignored and taken from the climindvis_index object. In case of autoplot functions, these values are handed over automatically and th_object should be left at default value of NULL..}

\item{q_threshold}{Percentile threshold (numerical). Value between 0 and 100.}

\item{NAmaxbasep}{Maximum number of NA values in aggregation period for all years (as defined by baseperiod). If exceeded, quantile value will be set to NA. Default = 20}

\item{qens_all=TRUE}{logical. Should quantiles of ensembles be calculated over all ensemble members (TRUE) or for each member individually (FALSE)? Default = TRUE}
}
\description{
quantile
}
\keyword{internal}
