% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_get_arguments.R
\name{th_doc}
\alias{th_doc}
\title{threshold}
\arguments{
\item{threshold}{Threshold (numerical). Thresholds need to be provided in mm for precipitation and in degrees C for temperature.}

\item{operator}{One of ">",">=","<","<="}

\item{threshold2}{Threshold (numerical)}

\item{operator2}{One of ">",">=","<","<="}

\item{thvar}{Variable for which days in range should be calculated (tmin/tmax/tavg/prec) by threshold operator thvar operator2 threshold2}

\item{dd_threshold}{Precipitation threshold for dry days in mm. Default = 1}
}
\description{
threshold
}
\keyword{internal}
