% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_get_arguments.R
\name{varall}
\alias{varall}
\title{varall}
\arguments{
\item{climindvis}{Object of type climindvis as returned by \code{\link{make_object}} which must contain the variable "var".}
}
\description{
varall
}
\keyword{internal}
