% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_get_arguments.R
\name{vartavg}
\alias{vartavg}
\title{vartavg}
\arguments{
\item{climindvis}{Object of type climindvis as returned by \code{\link{make_object}} which must contain the variable average temperature (tavg)}
}
\description{
vartavg
}
\keyword{internal}
