% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices_get_arguments.R
\name{vartmin}
\alias{vartmin}
\title{vartmin}
\arguments{
\item{climindvis}{Object of type climindvis as returned by \code{\link{make_object}} which must contain the variable minimum temperature (tmin)}
}
\description{
vartmin
}
\keyword{internal}
