% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_documentation.R
\name{veri_args}
\alias{veri_args}
\title{veri_args}
\arguments{
\item{nmin}{Number of forecasts that are not-missing for forecast to be evaluated, otherwise NA is returned see \code{\link{easyVerification}{veriApply}}. Default=20.}

\item{na_rm}{Logical, should incomplete forecasts-observation pairs be used? see \code{\link{easyVerification}{veriApply}}. Default=TRUE}

\item{ncat}{Number of equidistant categories for converting continous forecast into category forecasts (only for category skill scores). Default=3}

\item{prob}{Optional parameters passed to \code{\link{easyVerification}{veriApply}} for user defined probability thresholds for converting continuous forecasts to category forecasts(only for category skill scores). Default=NULL (in this case prob is set to (1:(ncat-1))/ncat).}
}
\description{
optional arguments for verification function
}
