% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_documentation.R
\name{veri_doc}
\alias{veri_doc}
\title{veri docu}
\arguments{
\item{ncat}{Number of equidistant categories to divide forecast into (if probs is specified not used). Default=3 (=tercile forecasts).}

\item{prob}{Probability thresholds for converting continuous forecasts to category forecasts. If not provided then ncat equidistant categories are calulated: probs= 1:(ncat-1)/ncat.}

\item{cat_names}{Names of forecast categorys to be used for plot. Default = paste0("cat",1:ncat)}

\item{skillmin}{Minimum skill limit for which forecasts are shown. Default=NULL (no skill criterion applied).}

\item{veri_metric}{Skill metric to calculate, see \code{\link{verify_index}}.}
}
\description{
veri docu
}
\keyword{internal}
