% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_index.R
\name{verify_index}
\alias{verify_index}
\title{Function for verification of hindcasts of indices}
\usage{
verify_index(obs, hc, hc_ref = NULL, index, index_args = list(),
  selyears = NULL, veri_metrics, veri_args = list())
}
\arguments{
\item{obs}{Observations. "climindvis" object of type p/grid as output from \code{\link{make_object}} by using daily observational or Re-analysis data.}

\item{hc}{Hindcasts. "climindvis_index" object  of type hc_p/hc_grid as output from \code{\link{make_object}} using an ensemble of  daily seasonal hindcasts.}

\item{hc_ref}{Reference forecast for calculation of skill scores. "climindvis"  object of type hc_p/hc_grid as output from \code{\link{make_object}} sing an ensemble of  daily seasonal hindcasts .  If hc_ref=NULL climatology will be taken as reference, Default = NULL.}

\item{index}{Name of index to be calculated.}

\item{index_args}{List of arguments for index. See \code{\link{calc_index}}.}

\item{selyears}{Integer array of years to be selected from data. For indices with quantiles, years as defined in index_args are used for quantile calculation, but only selected years are plotted. For functions using seasonal forecast data, selyears is only applied to hindcasts.}

\item{veri_metrics}{Character array of verification metrics as defined in the \code{\link{easyVerification}} package, e.g. c("EnsCorr","EnsRpss"). Also see "Verification metrics" below.}

\item{veri_args}{Optional list of additional arguments for verification, see \code{\link{veri_args}}}
}
\value{
Object of class \emph{index} and "climindvis_verification" with following entries:
\itemize{
\item verification: Named list with verification outputs for each metric in \emph{veri_metric}.
\item veri_info: Information needed for plotting of verification results, such as limits and dimensions of metric
\item years: years used for verification
\item lon: Array of longitudes
\item lat: Array of latitudes
\item {data_info$obs and $hc ($hc_ref):}{Data_info taken from obs and hc input objects, e.g. name of dataset}
}
}
\description{
Verifies daily seasonal hindcasts (re-forecasts) of indices for gridded or station data. It is based on the packages \code{\link{easyVerification}} and \code{\link{SpecsVerification}}.
}
\section{Cateogry forecasts}{


For category skill scores (e.g. ROC area and Rps(s)) forecasts have to be provided in categories. The easyVerification package automatically converts continuous forecast to category forecasts. The optional parameter \emph{veri_args$ncat}  determines the number of equidistant forecast categories (e.g. 3:terciles, 4:quartiles, 5:quintiles). This corresponds to setting the variable \emph{prob} of \code{\link{easyVerification}{veriApply}} to prob=c(1:((ncat-1)/ncat)). However, if you want to use other probability thresholds you can use the argument \emph{veri_args$prob}. The use of absolute thresholds (as possible in \code{\link{easyVerification}{veriApply}}) is currently not implemented.
}

\section{Verification metrics}{


The verify_index function can be used of calculating all of the skill metrics defined in the packages \code{\link{easyVerification}}
and \code{\link{SpecsVerification}}. Skill scores ending in "ss" denote the corresponding skill scores (see below).
Currently the following verification metrics from the \code{\link{easyVerification}} package can be calculated (the ending "ss" denotes the corresponding skill scores):
\describe{
 \item{Ens2AFC}{Generalized Discrimination Score}
 \item{EnsCorr}{Correlation with Ensemble Mean}
 \item{EnsMae/EnsMaess}{Ensemble mean absolute error}
 \item{EnsMse/EnsMsess}{Ensemble mean squared error}
 \item{EnsRmse/EnsRmsess}{Square root of ensemble mean squared error}
 \item{EnsRoca/EnsRocss}{ Area Under the ROC Curve}
 \item{EnsSprErr}{Spread to Error Ratio}
 \item{FairSprErr}{ Fair Spread to Error Ratio}
 \item{EnsRps/EnsRpss}{ Ranked probability skill score}
 \item{FairRps/FairRpss}{Fair Ranked probability skill score}
 \item{EnsCrps/EnsCrpss}{Continuous Ranked probability skill score}
 \item{FairCrps/FairCrpss}{Fair Continuous Ranked probability skill score}
}
More information about the skill scores can be found in the respective package.

Skill scores are defined as (mean score - mean score for reference) / (perfect score - mean score for reference).  The skill score is zero if the mean score of the forecast equals the mean score of the reference forecast, and equals one if the mean score of the forecast equals the best possible score. If hc_ref is not defined, the climatology is taken as reference.

For more information about Fair scores, see e.g. Ferro, C. A. T. (2014). Fair scores for ensemble forecasts. Quarterly Journal of the Royal Meteorological Society, 140(683), 1917-1923.
}

\examples{

## load example hindcast and observational data for stations
## (see \\code{\\link{example_climindvis_objects}}):

data("object_hc_st", "object_st")

## verify hindcasts of indices using the skill metrics
## "EnsRmsess", "EnsRpss", "EnsCorr":

veri<-verify_index(obs = object_st, hc = object_hc_st, index="txx", index_args=list(aggt="seasonal") , veri_metrics = c("EnsRmsess", "EnsRpss", "EnsCorr"))


}
\seealso{
\code{\link{easyVerification}},  \code{\link{calc_index}}
}
