% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_index_spei.R
\name{calc_spi}
\alias{calc_spi}
\title{spi calculations}
\usage{
calc_spi(temp = NULL, date_factor, forecast = FALSE, timescale, ref,
  distribution, NAmaxAgg = 20, iformat, limit = 4)
}
\arguments{
\item{temp}{vector of daily precipitation data to calculate SPI}

\item{date_factor}{factors of dates}

\item{forecast}{logical; Returns SPI parameters of historical timeseries to transform forecast data. Default = FALSE}

\item{timescale}{Integer. Timescale for monthly SPI calculation, i.e.running mean for 1-monthly,3-monthly, etc. period. Default = 6.}

\item{ref}{Vector with start and end year for reference period. If no reference period is specified, the full period is used as reference period.}

\item{distribution}{A character string naming a distribution of package stats (see for example GammaDist). For more information on the choice of the distribution, see section Details. Default = "gamma".}

\item{NAmaxAgg}{number of days to check for <<mcdd>>}

\item{limit}{Truncate SPI values that are larger than a given threshold, as values larger than 4 are not reasonable. Default = 4. The truncation can be disabled by setting limit = Inf.}
}
\description{
spi calculations
}
\keyword{internal}
